mm_drift_detector/requirements.txt
streamlit==1.38.0
pandas>=2.1.0
numpy>=1.24.0
scipy>=1.10.0
pymupdf==1.24.9
pillow>=10.2.0
# Optional OCR: requires system tesseract binary installed
pytesseract>=0.3.10
scikit-learn>=1.3.0
imbalanced-learn>=0.12.3
faker>=25.2.0
matplotlib>=3.8.0
